requirements.txt
reportlab
Faker
PyMuPDF
groq
pydantic
python-dotenv
scikit-learn
dill
dnc
rdkit
streamlit